make skylakex sgemm code more friendly for readers
BTW some kernels were adjusted to improve performance
